Conflicts:
	MineSweeper/MineSweeper/MyController.m

+++ MineSweeper/MineSweeper/MyController.m
@@ -11,12 +11,12 @@
 @implementation MyController
 
 - (IBAction)clicked:(id)sender {
-  int r = [sender selectedRow];
-  int c = [sender selectedColumn];
+  NSInteger r = [sender selectedRow];
+  NSInteger c = [sender selectedColumn];
   NSButtonCell *bcell = [sender selectedCell];
   Cell *cell = [minefield cellAtRow:r Col:c];
   
-  NSLog(@"clicked: sender=%@, selected cell=%@ at row=%d, col=%d", sender, bcell, r, c);
+  NSLog(@"clicked: sender=%@, selected cell=%@ at row=%ld, col=%ld", sender, bcell, (long)r, (long)c);
   
   if (!cell.marked) {
     int result = [minefield exposeCellAtRow:r Col:c];
@@ -39,8 +39,8 @@
 }
 
 - (IBAction)levelSelect:(id)sender {
-  int index = [sender indexOfSelectedItem];
-  NSLog(@"levelSelected:sender=%@, index=%d", sender, index);
+  NSInteger index = [sender indexOfSelectedItem];
+  NSLog(@"levelSelected:sender=%@, index=%ld", sender, (long)index);
 }
 
 -(void)awakeFromNib {
